hotfix for deep pagination of facets
This is a quick solution that will work to prevent misuse. Since the
normal interaction context is a modal, it's not ideal to redirect to
root here. And it'd be nice to cap the backing offset, and not hardcode
the facet.page variable name here. But this will be solid for now.