allow SUBJECT_CORRELATION feature to be switched on/off via ENV
this is a provisional commit, potentially making this eligible
to be merged into "master" branch before being ready to be fully
enabled. This will make it easier to keep these long-running
branches in sync without merge conflicts, at the expense of having
some extra unused code in the main branch.